feat: reset password and display error message
