+++ apps/bilan-carbone/src/components/auth/NewPasswordForm.tsx
@@ -3,18 +3,10 @@
 import { useServerFunction } from '@/hooks/useServerFunction'
 import { getEnvRoute } from '@/services/email/utils'
 import { resetPassword } from '@/services/serverFunctions/user'
-import { EmailCommand, EmailCommandValidation } from '@/services/serverFunctions/user.command'
-import Form from '@abc-transitionbascarbone/components/src/base/Form'
-import LoadingButton from '@abc-transitionbascarbone/components/src/base/LoadingButton'
-import { FormTextField } from '@abc-transitionbascarbone/components/src/form/TextField'
+import NewPasswordFormCommon from '@abc-transitionbascarbone/components/src/auth/NewPasswordFormCommon'
 import { Environment } from '@abc-transitionbascarbone/db-common/enums'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { FormControl } from '@mui/material'
 import { useTranslations } from 'next-intl'
-import { useRouter, useSearchParams } from 'next/navigation'
-import { useEffect, useState } from 'react'
-import { useForm } from 'react-hook-form'
-import authStyles from './Auth.module.css'
+import { useRouter } from 'next/navigation'
 
 interface Props {
   environment?: Environment
@@ -22,60 +14,21 @@ interface Props {
 
 const NewPasswordForm = ({ environment = Environment.BC }: Props) => {
   const t = useTranslations('login.form')
-  const [submitting, setSubmitting] = useState(false)
-  const searchParams = useSearchParams()
   const { callServerFunction } = useServerFunction()
   const router = useRouter()
 
-  const { control, getValues, handleSubmit, setValue } = useForm<EmailCommand>({
-    resolver: zodResolver(EmailCommandValidation),
-    mode: 'onBlur',
-    reValidateMode: 'onChange',
-    defaultValues: {
-      email: searchParams.get('email') ?? '',
-    },
-  })
-
   const loginLink = getEnvRoute('login', environment)
 
-  const onSubmit = async () => {
-    setSubmitting(true)
-
-    await callServerFunction(() => resetPassword(getValues().email.toLowerCase(), environment), {
+  const resetPasswordHandler = async (email: string) => {
+    callServerFunction(() => resetPassword(email.toLowerCase(), environment), {
       getSuccessMessage: () => t('emailSent'),
       getErrorMessage: (error) => t(error),
       onSuccess: () => {
         router.push(loginLink)
       },
     })
-    setSubmitting(false)
   }
-
-  useEffect(() => {
-    const email = searchParams.get('email')
-    if (email) {
-      setValue('email', email)
-    }
-  }, [searchParams, setValue])
-
-  return (
-    <Form onSubmit={handleSubmit(onSubmit)} className="grow justify-center">
-      <FormControl className={authStyles.form}>
-        <FormTextField
-          control={control}
-          className={authStyles.input}
-          label={t('email')}
-          placeholder={t('emailPlaceholder')}
-          name="email"
-          data-testid="input-email"
-          trim
-        />
-        <LoadingButton type="submit" data-testid="reset-button" loading={submitting} fullWidth>
-          {t('reset')}
-        </LoadingButton>
-      </FormControl>
-    </Form>
-  )
+  return <NewPasswordFormCommon resetPassword={resetPasswordHandler} />
 }
 
 export default NewPasswordForm
