+++ apps/bilan-carbone/src/services/serverFunctions/user.command.ts
@@ -49,15 +49,6 @@ export const EditSettingsCommandValidation = z.object({
 
 export type EditSettingsCommand = z.infer<typeof EditSettingsCommandValidation>
 
-export const EmailCommandValidation = z.object({
-  email: z
-    .email()
-    .trim()
-    .transform((email) => email.toLowerCase()),
-})
-
-export type EmailCommand = z.infer<typeof EmailCommandValidation>
-
 export const ResetPasswordCommandValidation = z.object({
   email: z
     .email()
