+++ packages/components/src/auth/user.command.ts
@@ -5,3 +5,12 @@ export const LoginCommandValidation = z.object({
 })
 
 export type LoginCommand = z.infer<typeof LoginCommandValidation>
+
+export const EmailCommandValidation = z.object({
+  email: z
+    .email()
+    .trim()
+    .transform((email) => email.toLowerCase()),
+})
+
+export type EmailCommand = z.infer<typeof EmailCommandValidation>
